+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+
+- Chart on event page, day comparative
+
 ## [0.0.33] - 2026-05-16
 
 ### Add